Add support for not using temp directory.

+++ sea.m
@@ -2,11 +2,14 @@ function sea(archive,varargin)
     expected_rename = {'none','partial','full'};
 
     p = inputParser;
+    p.KeepUnmatched = true;
     addRequired(p,'input');
     addParameter(p,'output','',@ischar);
     addParameter(p,'remove_temp',true);   
+    addParameter(p,'use_tempdir',true);   
     addParameter(p,'var1','a'); 
     addParameter(p,'var2','r'); 
+    addParameter(p,'main','z'); 
     addParameter(p,'cache',true); 
     addParameter(p,'rename_support','full',@(x) any(validatestring(x,expected_rename)))
     parse(p,archive,varargin{:});
@@ -29,21 +32,33 @@ function sea(archive,varargin)
     end
     outputfile = fullfile(outputpath,[outputname outputext]);
     
-    f = '%1$s=tempname,%2$s=unzip(%3$s,%1$s),run(%2$s{1})';
-    
-    if p.Results.remove_temp
-        f = [f ',rmdir(%1$s,''s'')'];
-    end
-    
-    if strcmp(p.Results.rename_support,'full')
-        namecmd = 'which(mfilename)';
-    elseif strcmp(p.Results.rename_support,'partial')
-        namecmd = '[mfilename,''.p'']';
+    if p.Results.use_tempdir
+        f = '%1$s=tempname,%2$s=unzip(%3$s,%1$s),run(%2$s{1})';
+        if p.Results.remove_temp
+            f = [f ',rmdir(%1$s,''s'')'];
+        end
+        if strcmp(p.Results.rename_support,'full')
+            namecmd = 'which(mfilename)';
+        elseif strcmp(p.Results.rename_support,'partial')
+            namecmd = '[mfilename,''.p'']';
+        else
+            namecmd = ['''' outputname outputext ''''];
+        end
     else
-        namecmd = ['''' outputname outputext ''''];
+        f = 'unzip%3$s;run %4$s';
+        if p.Results.remove_temp
+            f = [f ';delete %4$s.m'];
+        end
+        if strcmp(p.Results.rename_support,'full')
+            namecmd = '(which(mfilename))';
+        elseif strcmp(p.Results.rename_support,'partial')
+            namecmd = '([mfilename,''.p''])';
+        else
+            namecmd = [' ' outputname outputext ''];
+        end
     end
-    
-    scr = sprintf(f,p.Results.var1,p.Results.var2,namecmd);
+
+    scr = sprintf(f,p.Results.var1,p.Results.var2,namecmd,p.Results.main);
     
     scriptfile = ['s' num2str(hash(scr))];
